refactor llrt:util
