+++ libs/llrt_logging/Cargo.toml
@@ -21,4 +21,3 @@ rquickjs = { git = "https://github.com/DelSkayn/rquickjs.git", version = "0.9.0"
   "macro",
 ], default-features = false }
 ryu = "1"
-terminal_size = "0.4"
