+++ llrt_core/Cargo.toml
@@ -54,6 +54,7 @@ ryu = "1"
 simd-json = { version = "0.14", default-features = false, features = [
   "big-int-as-float",
 ] }
+terminal_size = "0.4"
 tokio = { version = "1", features = ["full"] }
 tracing = { version = "0.1", features = ["log"] }
 url = "=2.5.1"
